LevelChecker, Win State, Pause
Added a script/object to Hub to check if a level is marked complete in PublicVars. If it is, it will place a gold star over the corresponding portal. NOTE: If you change a scene's name, you must edit the Portal script to ensure the level is tracked as completed. More details in discord.

Setup the framework for a winstate. Basically if all golden stars are appeared when you go to the hub, right now it loads you to an endless state of loading to a "mirror" hub. I'll either develop a whole scene for the winstate or just make a text prompt.

Bundle the pause ui, event handler and game manager into the "ImportantBundle" prefab and added it to each level. Slightly modified it to get it to look better/work for different aspect ratios.